+++ server/routes/locations.ts
@@ -2,11 +2,13 @@ import express from "express"
 
 import * as db from '../db/db_functions/locations'
 import { validateAccessToken } from "../auth0"
+import checkAdmin from "./checkAdmin"
 
 const router = express.Router()
 
 router.get('/', validateAccessToken, async (req, res) => {
   try {
+    checkAdmin(req)
     const response = await db.getAllLocations()
     res.status(200).json(response)
   } catch (error) {
@@ -17,6 +19,7 @@ router.get('/', validateAccessToken, async (req, res) => {
 
 router.delete('/:locationId', validateAccessToken, async (req, res) => {
   try {
+    checkAdmin(req)
     const locationId = Number(req.params.locationId)
     await db.deleteLocation(locationId)
 
@@ -29,6 +32,7 @@ router.delete('/:locationId', validateAccessToken, async (req, res) => {
 
 router.post('/', validateAccessToken, async (req, res) => {
   try {
+    checkAdmin(req)
      const newLocation = req.body
      await db.addLocation(newLocation)
 
@@ -41,6 +45,7 @@ router.post('/', validateAccessToken, async (req, res) => {
 
 router.put('/:locationId', validateAccessToken, async (req, res) => {
   try {
+    checkAdmin(req)
     const locationId = Number(req.params.locationId)
     const updatedLocation = req.body
     await db.editLocation(locationId, updatedLocation)
